mysite/requirements.txt
asgiref==3.7.2
Django==4.2.6
packaging==23.2
Pillow==10.1.0
sqlparse==0.4.4
typing-extensions==4.8.0
tzdata==2023.3
waitress==2.1.2
